refactor: move supported languages to a separate static module
