+++ src/static/index.ts
@@ -1,0 +1,20 @@
+const supportedLanguages = [
+  'css',
+  'graphql',
+  'html',
+  'javascript',
+  'javascriptreact',
+  'json',
+  'jsonc',
+  'less',
+  'markdown',
+  'mdx',
+  'scss',
+  'svelte',
+  'typescript',
+  'typescriptreact',
+  'vue',
+  'yaml',
+];
+
+export = supportedLanguages;
